🐛 중앙 스토어 도입으로 데이터 불일치 해결
